requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.2
python-dotenv==1.0.0
werkzeug<3.0.0
flask-wtf==1.1.1
wtforms==3.0.1